Separate Cassandra persistence related structs * Separate shard related logic into dedicated struct * Separate mutable state related logic into dedicated struct * Separate mutable state task related logic into dedicated struct * Separate history related logic into dedicated struct * Rename files
