+++ common/persistence/cassandra/cluster_metadata.go
@@ -63,34 +63,26 @@ WHERE membership_partition = ?`
 )
 
 type (
-	cassandraClusterMetadata struct {
-		logger log.Logger
-		*cassandraStore
+	ClusterMetadataPersistence struct {
+		session gocql.Session
+		logger  log.Logger
 	}
 )
 
-var _ p.ClusterMetadataStore = (*cassandraClusterMetadata)(nil)
+var _ p.ClusterMetadataStore = (*ClusterMetadataPersistence)(nil)
 
 // newClusterMetadataInstance is used to create an instance of ClusterMetadataStore implementation
 func newClusterMetadataInstance(
 	session gocql.Session,
 	logger log.Logger,
 ) (p.ClusterMetadataStore, error) {
-
-	return &cassandraClusterMetadata{
-		cassandraStore: &cassandraStore{session: session, logger: logger},
-		logger:         logger,
+	return &ClusterMetadataPersistence{
+		session: session,
+		logger:  logger,
 	}, nil
 }
 
-// Close releases the resources held by this object
-func (m *cassandraClusterMetadata) Close() {
-	if m.session != nil {
-		m.session.Close()
-	}
-}
-
-func (m *cassandraClusterMetadata) GetClusterMetadata() (*p.InternalGetClusterMetadataResponse, error) {
+func (m *ClusterMetadataPersistence) GetClusterMetadata() (*p.InternalGetClusterMetadataResponse, error) {
 	query := m.session.Query(templateGetClusterMetadata, constMetadataPartition)
 	var clusterMetadata []byte
 	var encoding string
@@ -106,7 +98,7 @@ func (m *cassandraClusterMetadata) GetClusterMetadata() (*p.InternalGetClusterMe
 	}, nil
 }
 
-func (m *cassandraClusterMetadata) SaveClusterMetadata(request *p.InternalSaveClusterMetadataRequest) (bool, error) {
+func (m *ClusterMetadataPersistence) SaveClusterMetadata(request *p.InternalSaveClusterMetadataRequest) (bool, error) {
 	var query gocql.Query
 	if request.Version == 0 {
 		query = m.session.Query(
@@ -138,7 +130,7 @@ func (m *cassandraClusterMetadata) SaveClusterMetadata(request *p.InternalSaveCl
 	return true, nil
 }
 
-func (m *cassandraClusterMetadata) GetClusterMembers(request *p.GetClusterMembersRequest) (*p.GetClusterMembersResponse, error) {
+func (m *ClusterMetadataPersistence) GetClusterMembers(request *p.GetClusterMembersRequest) (*p.GetClusterMembersResponse, error) {
 	var queryString strings.Builder
 	var operands []interface{}
 	queryString.WriteString(templateGetClusterMembership)
@@ -212,7 +204,7 @@ func (m *cassandraClusterMetadata) GetClusterMembers(request *p.GetClusterMember
 	return &p.GetClusterMembersResponse{ActiveMembers: clusterMembers, NextPageToken: pagingToken}, nil
 }
 
-func (m *cassandraClusterMetadata) UpsertClusterMembership(request *p.UpsertClusterMembershipRequest) error {
+func (m *ClusterMetadataPersistence) UpsertClusterMembership(request *p.UpsertClusterMembershipRequest) error {
 	query := m.session.Query(templateUpsertActiveClusterMembership, constMembershipPartition, []byte(request.HostID),
 		request.RPCAddress, request.RPCPort, request.Role, request.SessionStart, time.Now().UTC(), int64(request.RecordExpiry.Seconds()))
 	err := query.Exec()
@@ -224,6 +216,16 @@ func (m *cassandraClusterMetadata) UpsertClusterMembership(request *p.UpsertClus
 	return nil
 }
 
-func (m *cassandraClusterMetadata) PruneClusterMembership(request *p.PruneClusterMembershipRequest) error {
+func (m *ClusterMetadataPersistence) PruneClusterMembership(request *p.PruneClusterMembershipRequest) error {
 	return nil
+}
+
+func (m *ClusterMetadataPersistence) GetName() string {
+	return cassandraPersistenceName
+}
+
+func (m *ClusterMetadataPersistence) Close() {
+	if m.session != nil {
+		m.session.Close()
+	}
 }
