+++ common/persistence/cassandra/execution_persistence.go
@@ -1,0 +1,192 @@
+// The MIT License
+//
+// Copyright (c) 2020 Temporal Technologies Inc.  All rights reserved.
+//
+// Copyright (c) 2020 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package cassandra
+
+import (
+	"time"
+
+	"go.temporal.io/server/common/log"
+	p "go.temporal.io/server/common/persistence"
+	"go.temporal.io/server/common/persistence/nosql/nosqlplugin/cassandra/gocql"
+)
+
+//	"go.temporal.io/api/serviceerror"
+// Guidelines for creating new special UUID constants
+// Each UUID should be of the form: E0000000-R000-f000-f000-00000000000x
+// Where x is any hexadecimal value, E represents the entity type valid values are:
+// E = {NamespaceID = 1, WorkflowID = 2, RunID = 3}
+// R represents row type in executions table, valid values are:
+// R = {Shard = 1, Execution = 2, Transfer = 3, Timer = 4, Replication = 5}
+const (
+	// Special Namespaces related constants
+	emptyNamespaceID = "10000000-0000-f000-f000-000000000000"
+	// Special Run IDs
+	emptyRunID     = "30000000-0000-f000-f000-000000000000"
+	permanentRunID = "30000000-0000-f000-f000-000000000001"
+	// Row Constants for Shard Row
+	rowTypeShardNamespaceID = "10000000-1000-f000-f000-000000000000"
+	rowTypeShardWorkflowID  = "20000000-1000-f000-f000-000000000000"
+	rowTypeShardRunID       = "30000000-1000-f000-f000-000000000000"
+	// Row Constants for Transfer Task Row
+	rowTypeTransferNamespaceID = "10000000-3000-f000-f000-000000000000"
+	rowTypeTransferWorkflowID  = "20000000-3000-f000-f000-000000000000"
+	rowTypeTransferRunID       = "30000000-3000-f000-f000-000000000000"
+	// Row Constants for Timer Task Row
+	rowTypeTimerNamespaceID = "10000000-4000-f000-f000-000000000000"
+	rowTypeTimerWorkflowID  = "20000000-4000-f000-f000-000000000000"
+	rowTypeTimerRunID       = "30000000-4000-f000-f000-000000000000"
+	// Row Constants for Replication Task Row
+	rowTypeReplicationNamespaceID = "10000000-5000-f000-f000-000000000000"
+	rowTypeReplicationWorkflowID  = "20000000-5000-f000-f000-000000000000"
+	rowTypeReplicationRunID       = "30000000-5000-f000-f000-000000000000"
+	// Row constants for visibility task row.
+	rowTypeVisibilityTaskNamespaceID = "10000000-6000-f000-f000-000000000000"
+	rowTypeVisibilityTaskWorkflowID  = "20000000-6000-f000-f000-000000000000"
+	rowTypeVisibilityTaskRunID       = "30000000-6000-f000-f000-000000000000"
+	// Row Constants for Replication Task DLQ Row. Source cluster name will be used as WorkflowID.
+	rowTypeDLQNamespaceID = "10000000-6000-f000-f000-000000000000"
+	rowTypeDLQRunID       = "30000000-6000-f000-f000-000000000000"
+	// Special TaskId constants
+	rowTypeExecutionTaskID = int64(-10)
+	rowTypeShardTaskID     = int64(-11)
+	emptyInitiatedID       = int64(-7)
+)
+
+const (
+	// Row types for table executions
+	rowTypeShard = iota
+	rowTypeExecution
+	rowTypeTransferTask
+	rowTypeTimerTask
+	rowTypeReplicationTask
+	rowTypeDLQ
+	rowTypeVisibilityTask
+)
+
+const (
+	// Row types for table tasks
+	rowTypeTask = iota
+	rowTypeTaskQueue
+)
+
+const (
+	taskQueueTaskID = -12345
+
+	// ref: https://docs.datastax.com/en/dse-trblshoot/doc/troubleshooting/recoveringTtlYear2038Problem.html
+	maxCassandraTTL = int64(315360000) // Cassandra max support time is 2038-01-19T03:14:06+00:00. Updated this to 10 years to support until year 2028
+)
+
+var (
+	defaultDateTime            = time.Date(2000, time.January, 1, 0, 0, 0, 0, time.UTC)
+	defaultVisibilityTimestamp = p.UnixMilliseconds(defaultDateTime)
+)
+
+type (
+	ExecutionPersistence struct {
+		*HistoryPersistence
+		*MutableStatePersistence
+		*MutableStateTaskPersistence
+	}
+)
+
+var _ p.ExecutionStore = (*ExecutionPersistence)(nil)
+
+func NewExecutionPersistence(
+	session gocql.Session,
+	logger log.Logger,
+) *ExecutionPersistence {
+	return &ExecutionPersistence{
+		HistoryPersistence:          NewHistoryPersistence(session, logger),
+		MutableStatePersistence:     NewMutableStatePersistence(session, logger),
+		MutableStateTaskPersistence: NewMutableStateTaskPersistence(session, logger),
+	}
+}
+
+func (d *ExecutionPersistence) CreateWorkflowExecution(
+	request *p.InternalCreateWorkflowExecutionRequest,
+) (*p.InternalCreateWorkflowExecutionResponse, error) {
+	for _, req := range request.NewWorkflowNewEvents {
+		if err := d.AppendHistoryNodes(req); err != nil {
+			return nil, err
+		}
+	}
+
+	return d.MutableStatePersistence.CreateWorkflowExecution(request)
+}
+
+func (d *ExecutionPersistence) UpdateWorkflowExecution(
+	request *p.InternalUpdateWorkflowExecutionRequest,
+) error {
+	for _, req := range request.UpdateWorkflowNewEvents {
+		if err := d.AppendHistoryNodes(req); err != nil {
+			return err
+		}
+	}
+	for _, req := range request.NewWorkflowNewEvents {
+		if err := d.AppendHistoryNodes(req); err != nil {
+			return err
+		}
+	}
+
+	return d.MutableStatePersistence.UpdateWorkflowExecution(request)
+}
+
+func (d *ExecutionPersistence) ConflictResolveWorkflowExecution(
+	request *p.InternalConflictResolveWorkflowExecutionRequest,
+) error {
+	for _, req := range request.CurrentWorkflowEventsNewEvents {
+		if err := d.AppendHistoryNodes(req); err != nil {
+			return err
+		}
+	}
+	for _, req := range request.ResetWorkflowEventsNewEvents {
+		if err := d.AppendHistoryNodes(req); err != nil {
+			return err
+		}
+	}
+	for _, req := range request.NewWorkflowEventsNewEvents {
+		if err := d.AppendHistoryNodes(req); err != nil {
+			return err
+		}
+	}
+
+	return d.MutableStatePersistence.ConflictResolveWorkflowExecution(request)
+}
+
+func (d *ExecutionPersistence) GetName() string {
+	return cassandraPersistenceName
+}
+
+func (d *ExecutionPersistence) Close() {
+	if d.HistoryPersistence.Session != nil {
+		d.HistoryPersistence.Session.Close()
+	}
+	if d.MutableStatePersistence.Session != nil {
+		d.MutableStatePersistence.Session.Close()
+	}
+	if d.MutableStateTaskPersistence.Session != nil {
+		d.MutableStateTaskPersistence.Session.Close()
+	}
+}
