+++ common/persistence/cassandra/matching_task_persistence.go
@@ -1,0 +1,425 @@
+// The MIT License
+//
+// Copyright (c) 2020 Temporal Technologies Inc.  All rights reserved.
+//
+// Copyright (c) 2020 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package cassandra
+
+import (
+	"fmt"
+	"strings"
+	"time"
+
+	enumspb "go.temporal.io/api/enums/v1"
+	"go.temporal.io/api/serviceerror"
+
+	"go.temporal.io/server/common/convert"
+	"go.temporal.io/server/common/log"
+	p "go.temporal.io/server/common/persistence"
+	"go.temporal.io/server/common/persistence/nosql/nosqlplugin/cassandra/gocql"
+	"go.temporal.io/server/common/primitives/timestamp"
+)
+
+type (
+	MatchingTaskPersistence struct {
+		Session gocql.Session
+		Logger  log.Logger
+	}
+)
+
+func NewMatchingTaskPersistence(
+	session gocql.Session,
+	logger log.Logger,
+) *MatchingTaskPersistence {
+	return &MatchingTaskPersistence{
+		Session: session,
+		Logger:  logger,
+	}
+}
+
+func (d *MatchingTaskPersistence) CreateTaskQueue(
+	request *p.InternalCreateTaskQueueRequest,
+) error {
+	query := d.Session.Query(templateInsertTaskQueueQuery,
+		request.NamespaceID,
+		request.TaskQueue,
+		request.TaskType,
+		rowTypeTaskQueue,
+		taskQueueTaskID,
+		request.RangeID,
+		request.TaskQueueInfo.Data,
+		request.TaskQueueInfo.EncodingType.String(),
+	)
+
+	previous := make(map[string]interface{})
+	applied, err := query.MapScanCAS(previous)
+	if err != nil {
+		return gocql.ConvertError("LeaseTaskQueue", err)
+	}
+
+	if !applied {
+		previousRangeID := previous["range_id"]
+		return &p.ConditionFailedError{
+			Msg: fmt.Sprintf("CreateTaskQueue: TaskQueue:%v, TaskQueueType:%v, PreviousRangeID:%v",
+				request.TaskQueue, request.TaskType, previousRangeID),
+		}
+	}
+
+	return nil
+}
+
+func (d *MatchingTaskPersistence) GetTaskQueue(
+	request *p.InternalGetTaskQueueRequest,
+) (*p.InternalGetTaskQueueResponse, error) {
+	query := d.Session.Query(templateGetTaskQueue,
+		request.NamespaceID,
+		request.TaskQueue,
+		request.TaskType,
+		rowTypeTaskQueue,
+		taskQueueTaskID,
+	)
+
+	var rangeID int64
+	var tlBytes []byte
+	var tlEncoding string
+	if err := query.Scan(&rangeID, &tlBytes, &tlEncoding); err != nil {
+		return nil, gocql.ConvertError("GetTaskQueue", err)
+	}
+
+	return &p.InternalGetTaskQueueResponse{
+		RangeID:       rangeID,
+		TaskQueueInfo: p.NewDataBlob(tlBytes, tlEncoding),
+	}, nil
+}
+
+func (d *MatchingTaskPersistence) ExtendLease(
+	request *p.InternalExtendLeaseRequest,
+) error {
+	query := d.Session.Query(templateUpdateTaskQueueQuery,
+		request.RangeID+1,
+		request.TaskQueueInfo.Data,
+		request.TaskQueueInfo.EncodingType.String(),
+		request.NamespaceID,
+		&request.TaskQueue,
+		request.TaskType,
+		rowTypeTaskQueue,
+		taskQueueTaskID,
+		request.RangeID,
+	)
+	previous := make(map[string]interface{})
+	applied, err := query.MapScanCAS(previous)
+	if err != nil {
+		return gocql.ConvertError("LeaseTaskQueue", err)
+	}
+
+	if !applied {
+		previousRangeID := previous["range_id"]
+		return &p.ConditionFailedError{
+			Msg: fmt.Sprintf("ExtendLease: taskQueue:%v, taskQueueType:%v, haveRangeID:%v, gotRangeID:%v",
+				request.TaskQueue, request.TaskType, request.RangeID, previousRangeID),
+		}
+	}
+
+	return nil
+}
+
+// UpdateTaskQueue update task queue
+func (d *MatchingTaskPersistence) UpdateTaskQueue(
+	request *p.InternalUpdateTaskQueueRequest,
+) (*p.UpdateTaskQueueResponse, error) {
+	var err error
+	var applied bool
+	previous := make(map[string]interface{})
+	if request.TaskQueueKind == enumspb.TASK_QUEUE_KIND_STICKY { // if task_queue is sticky, then update with TTL
+		if request.ExpiryTime == nil {
+			return nil, serviceerror.NewInternal("ExpiryTime cannot be nil for sticky task queue")
+		}
+		expiryTtl := convert.Int64Ceil(time.Until(timestamp.TimeValue(request.ExpiryTime)).Seconds())
+		batch := d.Session.NewBatch(gocql.LoggedBatch)
+		batch.Query(templateUpdateTaskQueueQueryWithTTLPart1,
+			request.NamespaceID,
+			request.TaskQueue,
+			request.TaskType,
+			rowTypeTaskQueue,
+			taskQueueTaskID,
+			expiryTtl,
+		)
+		batch.Query(templateUpdateTaskQueueQueryWithTTLPart2,
+			expiryTtl,
+			request.RangeID,
+			request.TaskQueueInfo.Data,
+			request.TaskQueueInfo.EncodingType.String(),
+			request.NamespaceID,
+			request.TaskQueue,
+			request.TaskType,
+			rowTypeTaskQueue,
+			taskQueueTaskID,
+			request.RangeID,
+		)
+		applied, _, err = d.Session.MapExecuteBatchCAS(batch, previous)
+	} else {
+		query := d.Session.Query(templateUpdateTaskQueueQuery,
+			request.RangeID,
+			request.TaskQueueInfo.Data,
+			request.TaskQueueInfo.EncodingType.String(),
+			request.NamespaceID,
+			request.TaskQueue,
+			request.TaskType,
+			rowTypeTaskQueue,
+			taskQueueTaskID,
+			request.RangeID,
+		)
+		applied, err = query.MapScanCAS(previous)
+	}
+
+	if err != nil {
+		return nil, gocql.ConvertError("UpdateTaskQueue", err)
+	}
+
+	if !applied {
+		var columns []string
+		for k, v := range previous {
+			columns = append(columns, fmt.Sprintf("%s=%v", k, v))
+		}
+
+		return nil, &p.ConditionFailedError{
+			Msg: fmt.Sprintf("Failed to update task queue. name: %v, type: %v, rangeID: %v, columns: (%v)",
+				request.TaskQueue, request.TaskType, request.RangeID, strings.Join(columns, ",")),
+		}
+	}
+
+	return &p.UpdateTaskQueueResponse{}, nil
+}
+
+func (d *MatchingTaskPersistence) ListTaskQueue(
+	_ *p.ListTaskQueueRequest,
+) (*p.InternalListTaskQueueResponse, error) {
+	return nil, serviceerror.NewUnavailable(fmt.Sprintf("unsupported operation"))
+}
+
+func (d *MatchingTaskPersistence) DeleteTaskQueue(
+	request *p.DeleteTaskQueueRequest,
+) error {
+	query := d.Session.Query(templateDeleteTaskQueueQuery,
+		request.TaskQueue.NamespaceID, request.TaskQueue.Name, request.TaskQueue.TaskType, rowTypeTaskQueue, taskQueueTaskID, request.RangeID)
+	previous := make(map[string]interface{})
+	applied, err := query.MapScanCAS(previous)
+	if err != nil {
+		return gocql.ConvertError("DeleteTaskQueue", err)
+	}
+	if !applied {
+		return &p.ConditionFailedError{
+			Msg: fmt.Sprintf("DeleteTaskQueue operation failed: expected_range_id=%v but found %+v", request.RangeID, previous),
+		}
+	}
+	return nil
+}
+
+// CreateTasks add tasks
+func (d *MatchingTaskPersistence) CreateTasks(
+	request *p.InternalCreateTasksRequest,
+) (*p.CreateTasksResponse, error) {
+	batch := d.Session.NewBatch(gocql.LoggedBatch)
+	namespaceID := request.NamespaceID
+	taskQueue := request.TaskQueue
+	taskQueueType := request.TaskType
+
+	for _, task := range request.Tasks {
+		ttl := GetTaskTTL(task.ExpiryTime)
+
+		if ttl <= 0 || ttl > maxCassandraTTL {
+			batch.Query(templateCreateTaskQuery,
+				namespaceID,
+				taskQueue,
+				taskQueueType,
+				rowTypeTask,
+				task.TaskId,
+				task.Task.Data,
+				task.Task.EncodingType.String())
+		} else {
+			batch.Query(templateCreateTaskWithTTLQuery,
+				namespaceID,
+				taskQueue,
+				taskQueueType,
+				rowTypeTask,
+				task.TaskId,
+				task.Task.Data,
+				task.Task.EncodingType.String(),
+				ttl)
+		}
+	}
+
+	// The following query is used to ensure that range_id didn't change
+	batch.Query(templateUpdateTaskQueueQuery,
+		request.RangeID,
+		request.TaskQueueInfo.Data,
+		request.TaskQueueInfo.EncodingType.String(),
+		namespaceID,
+		taskQueue,
+		taskQueueType,
+		rowTypeTaskQueue,
+		taskQueueTaskID,
+		request.RangeID,
+	)
+
+	previous := make(map[string]interface{})
+	applied, _, err := d.Session.MapExecuteBatchCAS(batch, previous)
+	if err != nil {
+		return nil, gocql.ConvertError("CreateTasks", err)
+	}
+	if !applied {
+		rangeID := previous["range_id"]
+		return nil, &p.ConditionFailedError{
+			Msg: fmt.Sprintf("Failed to create task. TaskQueue: %v, taskQueueType: %v, rangeID: %v, db rangeID: %v",
+				taskQueue, taskQueueType, request.RangeID, rangeID),
+		}
+	}
+
+	return &p.CreateTasksResponse{}, nil
+}
+
+func GetTaskTTL(expireTime *time.Time) int64 {
+	var ttl int64 = 0
+	if expireTime != nil {
+		expiryTtl := convert.Int64Ceil(time.Until(timestamp.TimeValue(expireTime)).Seconds())
+
+		// 0 means no ttl, we dont want that.
+		// Todo: Come back and correctly ignore expired in-memory tasks before persisting
+		if expiryTtl < 1 {
+			expiryTtl = 1
+		}
+
+		ttl = expiryTtl
+	}
+	return ttl
+}
+
+// GetTasks get a task
+func (d *MatchingTaskPersistence) GetTasks(
+	request *p.GetTasksRequest,
+) (*p.InternalGetTasksResponse, error) {
+	if request.MaxReadLevel == nil {
+		return nil, serviceerror.NewInternal("getTasks: both readLevel and maxReadLevel MUST be specified for cassandra persistence")
+	}
+	if request.ReadLevel > *request.MaxReadLevel {
+		return &p.InternalGetTasksResponse{}, nil
+	}
+
+	// Reading taskqueue tasks need to be quorum level consistent, otherwise we could lose tasks
+	query := d.Session.Query(templateGetTasksQuery,
+		request.NamespaceID,
+		request.TaskQueue,
+		request.TaskType,
+		rowTypeTask,
+		request.ReadLevel,
+		*request.MaxReadLevel,
+	)
+	iter := query.PageSize(request.BatchSize).Iter()
+
+	response := &p.InternalGetTasksResponse{}
+	task := make(map[string]interface{})
+PopulateTasks:
+	for iter.MapScan(task) {
+		_, ok := task["task_id"]
+		if !ok { // no tasks, but static column record returned
+			continue
+		}
+
+		rawTask, ok := task["task"]
+		if !ok {
+			return nil, newFieldNotFoundError("task", task)
+		}
+		taskVal, ok := rawTask.([]byte)
+		if !ok {
+			var byteSliceType []byte
+			return nil, newPersistedTypeMismatchError("task", byteSliceType, rawTask, task)
+
+		}
+
+		rawEncoding, ok := task["task_encoding"]
+		if !ok {
+			return nil, newFieldNotFoundError("task_encoding", task)
+		}
+		encodingVal, ok := rawEncoding.(string)
+		if !ok {
+			var byteSliceType []byte
+			return nil, newPersistedTypeMismatchError("task_encoding", byteSliceType, rawEncoding, task)
+		}
+
+		response.Tasks = append(response.Tasks, p.NewDataBlob(taskVal, encodingVal))
+		if len(response.Tasks) == request.BatchSize {
+			break PopulateTasks
+		}
+		task = make(map[string]interface{}) // Reinitialize map as initialized fails on unmarshalling
+	}
+
+	if err := iter.Close(); err != nil {
+		return nil, serviceerror.NewUnavailable(fmt.Sprintf("GetTasks operation failed. Error: %v", err))
+	}
+
+	return response, nil
+}
+
+// CompleteTask delete a task
+func (d *MatchingTaskPersistence) CompleteTask(
+	request *p.CompleteTaskRequest,
+) error {
+	tli := request.TaskQueue
+	query := d.Session.Query(templateCompleteTaskQuery,
+		tli.NamespaceID,
+		tli.Name,
+		tli.TaskType,
+		rowTypeTask,
+		request.TaskID)
+
+	err := query.Exec()
+	if err != nil {
+		return gocql.ConvertError("CompleteTask", err)
+	}
+
+	return nil
+}
+
+// CompleteTasksLessThan deletes all tasks less than or equal to the given task id. This API ignores the
+// Limit request parameter i.e. either all tasks leq the task_id will be deleted or an error will
+// be returned to the caller
+func (d *MatchingTaskPersistence) CompleteTasksLessThan(
+	request *p.CompleteTasksLessThanRequest,
+) (int, error) {
+	query := d.Session.Query(templateCompleteTasksLessThanQuery,
+		request.NamespaceID, request.TaskQueueName, request.TaskType, rowTypeTask, request.TaskID)
+	err := query.Exec()
+	if err != nil {
+		return 0, gocql.ConvertError("CompleteTasksLessThan", err)
+	}
+	return p.UnknownNumRowsAffected, nil
+}
+
+func (d *MatchingTaskPersistence) GetName() string {
+	return cassandraPersistenceName
+}
+
+func (d *MatchingTaskPersistence) Close() {
+	if d.Session != nil {
+		d.Session.Close()
+	}
+}
